ファイルを単一関数のエクスポートに分割
