+++ lib/run.js
@@ -2,7 +2,8 @@ import process from "node:process";
 import path from "node:path";
 import { chromium } from "playwright";
 import envPaths from "env-paths";
-import { watchFile, watchDirectory } from "./watch.js";
+import { watchFile } from "./fileSystem/watchFile.js";
+import { watchDirectory } from "./fileSystem/watchDirectory.js";
 import { isNewViewPage, PreviewPage, getEnvironment } from "./PreviewPage.js";
 
 const appEnvPaths = envPaths("bm-view-preview");
